requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Babel==2.9.1
certifi==2022.12.7
charset-normalizer==3.0.1
docopt==0.6.2
frozenlist==1.3.3
h11==0.14.0
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.32
outcome==1.2.0
pipreqs==0.4.11
protobuf==3.20.3
PySocks==1.7.1
pytz==2022.7.1
requests==2.28.2
selenium==4.8.2
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.22.0
trio-websocket==0.9.2
urllib3==1.26.14
wsproto==1.2.0
yarg==0.1.9
yarl==1.8.2
